requirements.txt
streamlit>=1.36.0
crawl4ai>=0.4.0
openai>=1.40.0
python-dotenv>=1.0.1
pytest>=8.0.0